requirements.txt
numpy
pyqt5
pillow
gym
gym-retro